--- a/core/chain_makers.go
+++ b/core/chain_makers.go
@@ -53,8 +53,7 @@ type BlockGen struct {
 
 	config *params.ChainConfig
 	engine consensus.Engine
-
-	}
+}
 
 // SetCoinbase sets the coinbase of the generated block.
 // It can be called at most once.
@@ -100,18 +99,8 @@ func (b *BlockGen) AddTxWithChain(bc *BlockChain, tx *types.Transaction) {
 	}
 	b.statedb.Prepare(tx.Hash(), common.Hash{}, len(b.txs))
 
-	blockContext := &vm.BlockContext{
-		BlockNumber: b.header.Number,
-		Coinbase:    b.header.Coinbase,
-		GasLimit:    b.header.GasLimit,
-		Signer:      types.MakeSigner(b.config, b.header.Number),
-		Precompiles: vm.PrecompiledContractsByzantium, // TODO fix
-		Intpool:     vm.NewIntpool(),
-		Time:        b.header.Time,
-		Difficulty:  b.header.Difficulty,
-	}
-
-	receipt, _, err := ApplyTransaction(b.config, bc, &b.header.Coinbase, b.gasPool, b.statedb, b.header, tx, &b.header.GasUsed, &vm.Config{}, blockContext)
+	blockContext := NewBlockContext(b.header, b.header.Coinbase, b.config)
+	receipt, _, err := ApplyTransaction(b.config, bc, b.gasPool, b.statedb, b.header, tx, &b.header.GasUsed, &vm.Config{}, blockContext)
 	if err != nil {
 		panic(err)
 	}