--- a/core/evm.go
+++ b/core/evm.go
@@ -36,8 +36,8 @@ type ChainContext interface {
 }
 
 // NewEVMContext creates a new context for use in the EVM.
-func NewEVMContext(msg Message, header *types.Header, chain ChainContext, author *common.Address) vm.Context {
-	return vm.Context{
+func NewEVMContext(msg Message, header *types.Header, chain ChainContext, author *common.Address) *vm.Context {
+	return &vm.Context{
 		CanTransfer: CanTransfer,
 		Transfer:    Transfer,
 		GetHash:     GetHashFn(header, chain),