--- a/core/state_processor.go
+++ b/core/state_processor.go
@@ -25,6 +25,7 @@ import (
 	"github.com/ethereum/go-ethereum/core/vm"
 	"github.com/ethereum/go-ethereum/crypto"
 	"github.com/ethereum/go-ethereum/params"
+	"math/big"
 )
 
 // StateProcessor is a basic Processor, which takes care of transitioning
@@ -65,12 +66,29 @@ func (p *StateProcessor) Process(block *types.Block, statedb *state.StateDB, cfg
 	if p.config.DAOForkSupport && p.config.DAOForkBlock != nil && p.config.DAOForkBlock.Cmp(block.Number()) == 0 {
 		misc.ApplyDAOHardFork(statedb)
 	}
-	signer := types.MakeSigner(p.config, header.Number)
+
+	precompiles := vm.PrecompiledContractsHomestead
+	if p.config.ByzantiumBlock.Cmp(block.Number()) <= 0 {
+		precompiles = vm.PrecompiledContractsByzantium
+	}
+	beneficiary, _ := p.engine.Author(block.Header())
+
+	blockContext := &vm.BlockContext{
+		Intpool:     vm.NewIntpool(),
+		Precompiles: precompiles,
+
+		BlockNumber: new(big.Int).Set(block.Header().Number),
+		Time:        new(big.Int).Set(block.Header().Time),
+		Difficulty:  new(big.Int).Set(block.Header().Difficulty),
+		GasLimit:    block.Header().GasLimit,
+		Coinbase:    beneficiary,
+		Signer:      types.MakeSigner(p.config, block.Header().Number),
+	}
 	// Iterate over and process the individual transactions
 	receipts = make([]*types.Receipt, len(block.Transactions()))
 	for i, tx := range block.Transactions() {
 		statedb.Prepare(tx.Hash(), block.Hash(), i)
-		receipt, _, err := ApplyTransaction(p.config, p.bc, nil, gp, statedb, header, tx, usedGas, cfg, &signer)
+		receipt, _, err := ApplyTransaction(p.config, p.bc, nil, gp, statedb, header, tx, usedGas, cfg, blockContext)
 		if err != nil {
 			return nil, nil, 0, err
 		}
@@ -87,8 +105,8 @@ func (p *StateProcessor) Process(block *types.Block, statedb *state.StateDB, cfg
 // and uses the input parameters for its environment. It returns the receipt
 // for the transaction, gas used and an error if the transaction failed,
 // indicating the block was invalid.
-func ApplyTransaction(config *params.ChainConfig, bc *BlockChain, author *common.Address, gp *GasPool, statedb *state.StateDB, header *types.Header, tx *types.Transaction, usedGas *uint64, cfg vm.Config, signer *types.Signer) (*types.Receipt, uint64, error) {
-	msg, err := tx.AsMessage(*signer)
+func ApplyTransaction(config *params.ChainConfig, bc *BlockChain, author *common.Address, gp *GasPool, statedb *state.StateDB, header *types.Header, tx *types.Transaction, usedGas *uint64, cfg vm.Config, blockContext *vm.BlockContext) (*types.Receipt, uint64, error) {
+	msg, err := tx.AsMessage(blockContext.Signer)
 	if err != nil {
 		return nil, 0, err
 	}
@@ -96,7 +114,7 @@ func ApplyTransaction(config *params.ChainConfig, bc *BlockChain, author *common
 	context := NewEVMContext(msg, header, bc, author)
 	// Create a new environment which holds all relevant information
 	// about the transaction and calling mechanisms.
-	vmenv := vm.NewEVM(context, statedb, config, cfg)
+	vmenv := vm.NewEVM(context, statedb, config, cfg, blockContext)
 	// Apply the transaction to the current state (included in the env)
 	_, gas, failed, err := ApplyMessage(vmenv, msg, gp)
 	if err != nil {