--- a/core/vm/contract.go
+++ b/core/vm/contract.go
@@ -64,6 +64,7 @@ type Contract struct {
 	DelegateCall bool
 }
 
+// NewPrecompiledContract returns a new contract environment for the execution of a precompiled contract
 func NewPrecompiledContract( caller ContractRef,  object ContractRef,value *big.Int, gas uint64) *Contract{
 
 	c := &Contract{CallerAddress: caller.Address(), caller: caller, self: object, Args: nil}