--- a/core/vm/int_pool_verifier_empty.go
+++ b/core/vm/int_pool_verifier_empty.go
@@ -20,4 +20,4 @@ package vm
 
 const verifyPool = false
 
-func verifyIntegerPool(ip *intPool) {}
+func verifyIntegerPool(ip *IntPool) {}