This is a work in progress. Is __not__ ready for merging.

The benchmarks below are based on the following cases:

1. `BlockChain_1x1000ValueTransferToNonexisting` : importing one block, filled with `1000` transactions that each send a wei to a non-existing recipient. No data. 
2. `BlockChain_1x1000ValueTransferToExisting` : importing one block, filled with `1000` transactions that each send a wei to an _existing_ recipient. No data. 
3. `BlockChain_1x1000Executions`:  importing one block, filled with `1000` transactions that each invoke an existing contract which has code to do a `push` and a `pop`. No data. 

### Not allocating 1024 for intpool every tx

```
[user@work core]$ benchstat oldtimes.txt zerosizetimes.txt 
name                                           old time/op  new time/op  delta
BlockChain_1x1000ValueTransferToNonexisting-2  90.6ms ± 2%  85.5ms ± 5%   -5.54%  (p=0.032 n=5+5)
BlockChain_1x1000ValueTransferToExisting-2     62.8ms ± 6%  56.4ms ± 7%  -10.30%  (p=0.016 n=5+5)
BlockChain_1x1000Executions-2                  64.7ms ± 4%  60.7ms ± 4%   -6.08%  (p=0.008 n=5+5)
```
### Using a block-specific context

A per-block struct is used for things that does not change between transactions. This also hosts an intpool and the `precompile` list, among other things. Quite a large change. It has some benefits, but not extreme:
```
# Use block-context, which is not created anew each time
[user@work core]$ benchstat oldtimes.txt newtimes.txt 
name                                           old time/op  new time/op  delta
BlockChain_1x1000ValueTransferToNonexisting-2  90.6ms ± 2%  82.6ms ± 8%   -8.82%  (p=0.016 n=5+5)
BlockChain_1x1000ValueTransferToExisting-2     62.8ms ± 6%  56.0ms ± 7%  -10.81%  (p=0.008 n=5+5)
BlockChain_1x1000Executions-2                  64.7ms ± 4%  58.1ms ± 5%  -10.12%  (p=0.008 n=5+5)
```
### Pointers instead of objects

The last optimization is to replace large objects in the call chain for executing transactions, and using pointers instead. This change resulted in significant improvements. We need to ensure that at no point can anything mess with the data, though. 
```
# using pointers instead of passing around large structs
[user@work core]$ benchstat oldtimes.txt morepointertimes.txt 
name                                           old time/op  new time/op  delta
BlockChain_1x1000ValueTransferToNonexisting-2  90.6ms ± 2%  67.5ms ± 4%  -25.44%  (p=0.008 n=5+5)
BlockChain_1x1000ValueTransferToExisting-2     62.8ms ± 6%  42.1ms ± 4%  -33.02%  (p=0.008 n=5+5)
BlockChain_1x1000Executions-2                  64.7ms ± 4%  44.7ms ±12%  -30.89%  (p=0.008 n=5+5)
```

I'll post some graphs once I've run it on the aws cluster.