@holiman 2018-04-17
Experimental PR
Start:

```
Apr 12 11:37:24 18.130.0.187 geth-experimental: INFO [04-12|09:37:24] Block synchronisation started  
...

Apr 16 20:56:11 18.130.0.187 geth-experimental: INFO [04-16|18:56:11] Imported new chain segment               blocks=1    txs=91    mgas=5.239    elapsed=290.217ms mgasps=18.052  number=5452454 hash=720c1f…0f6496 cache=188.68mB 
```

From `0` to `5452454`: 105h 20m

Master:

```
Apr 12 11:37:51 35.177.24.226 geth-master: INFO [04-12|09:37:50] Block synchronisation started 
...

Apr 16 23:05:41 35.177.24.226 geth-master: INFO [04-16|21:05:41] Imported new chain segment               blocks=23   txs=2579  mgas=153.137 elapsed=8.036s    mgasps=19.056  number=5452460 hash=dbc5b1…be4f71 cache=207.04mB 
....
Apr 16 23:09:19 35.177.24.226 geth-master: INFO [04-16|21:09:19] Imported new chain segment               blocks=1    txs=31    mgas=3.046   elapsed=44.742ms  mgasps=68.079  number=5453015 hash=c8b164…65e97e cache=149.03mB 
```

From `0` to `5452460`: 107h 30m