@holiman 2018-05-28
![memstat_faster_blocks](https://user-images.githubusercontent.com/142290/40625234-83a20e1e-62b1-11e8-9980-bba5fbc1573e.png)
This graphs shows two nodes running, fast-synched and up to date. In the middle of the graph, one of the nodes was switched over to this PR. Although there was no immediate effect on block processing time, there was a drastic effect on the number of allocations performed, from ca 1M to 250K.