@GitCop 2018-05-29
Thank you for your contribution! Your commits seem to not adhere to the repository coding standards

 * Commit: 874ca182e3a8c61766529caaba8e0e669869d54c
  * Commits must be prefixed with the package(s) they modify

Please check the [contribution guidelines](https://github.com/ethereum/go-ethereum#contribution) for more details.

---
This message was auto-generated by https://gitcop.com