@holiman 2018-05-29
I have now re-enabled some stuff which I hack-disabled while testing out the PR. If travis turns green, it would be good to

1. Run this on a full import again
2. Verify that transaction tracing still works, as well as estimateGas
3. Verify that LES odr-stuff still works